requirements.txt
asgiref==3.8.1
Django==4.2.15
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.17
pytweening==1.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.11.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.41.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.2